Extract (a subset of) nexus::external_api into new nexus-types crate
