+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "nexus/db-macros",
     "nexus/test-utils",
     "nexus/test-utils-macros",
+    "nexus/types",
     "nexus-client",
     "package",
     "rpaths",
@@ -45,6 +46,7 @@ default-members = [
     "nexus",
     "nexus/authz-macros",
     "nexus/db-macros",
+    "nexus/types",
     "package",
     "rpaths",
     "sled-agent",
