+++ nexus/src/db/mod.rs
@@ -32,7 +32,6 @@ mod update_and_check;
 #[cfg(test)]
 mod test_utils;
 
-pub mod identity;
 pub mod model;
 pub mod schema;
 
@@ -42,3 +41,5 @@ pub use pool::Pool;
 pub use saga_recovery::{recover, RecoveryTask};
 pub use saga_types::SecId;
 pub use sec_store::CockroachDbSecStore;
+
+pub use nexus_types::identity;
