+++ nexus/types/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "nexus-types"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+
+[dependencies]
+anyhow = "1.0"
+chrono = { version = "0.4", features = ["serde"] }
+base64 = "0.13.0"
+# must match samael's crate!
+openssl = "0.10"
+openssl-sys = "0.9"
+openssl-probe = "0.1.2"
+schemars = { version = "0.8.10", features = ["chrono", "uuid1"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+uuid = { version = "1.1.0", features = ["serde", "v4"] }
+
+api_identity = { path = "../../api_identity" }
+omicron-common = { path = "../../common" }
