+++ nexus/types/src/external_api/mod.rs
@@ -2,10 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub mod console_api;
-pub mod device_auth;
-pub mod http_entrypoints;
-
-pub use nexus_types::external_api::params;
-pub use nexus_types::external_api::shared;
-pub use nexus_types::external_api::views;
+pub mod params;
+pub mod shared;
+pub mod views;
